+++ http4k-testing/chaos/src/test/kotlin/org/http4k/chaos/ChaosEngineTest.kt
@@ -3,7 +3,6 @@ package org.http4k.chaos
 import com.natpryce.hamkrest.and
 import com.natpryce.hamkrest.assertion.assertThat
 import com.natpryce.hamkrest.equalTo
-import org.http4k.chaos.ChaosBehaviours.NoBody
 import org.http4k.chaos.ChaosBehaviours.ReturnStatus
 import org.http4k.contract.security.ApiKeySecurity
 import org.http4k.core.Method.GET
@@ -49,11 +48,11 @@ class ChaosEngineTest {
 
         engine.enable()
         assertThat(appWithChaos(Request(GET, "/")), hasStatus(NOT_FOUND))
-//        assertThat(engine.toString(), equalTo("Always ReturnStatus (404)"))
+        assertThat(engine.toString(), equalTo("Always ReturnStatus (404)"))
 
         engine.enable(ReturnStatus(I_M_A_TEAPOT))
         assertThat(appWithChaos(Request(GET, "/")), hasStatus(I_M_A_TEAPOT))
-//        assertThat(engine.toString(), equalTo("Always ReturnStatus (418)"))
+        assertThat(engine.toString(), equalTo("Always ReturnStatus (418)"))
 
         engine.disable()
         assertThat(appWithChaos(Request(GET, "/")), hasStatus(OK))
@@ -70,14 +69,17 @@ class ChaosEngineTest {
 
         assertThat(appWithChaos(Request(GET, "/")), hasStatus(OK))
         assertThat(appWithChaos(Request(GET, "/chaos/status")), hasBody(noChaos))
-        assertThat(appWithChaos(Request(POST, "/chaos/activate")), hasStatus(OK))//.and(hasBody(originalChaos)))
-//        assertThat(appWithChaos(Request(GET, "/chaos/status")), hasBody(originalChaos))
+        assertThat(appWithChaos(Request(POST, "/chaos/activate")), hasStatus(OK).and(hasBody(originalChaos)))
+        assertThat(appWithChaos(Request(GET, "/chaos/status")), hasBody(originalChaos))
         assertThat(appWithChaos(Request(POST, "/")), hasStatus(METHOD_NOT_ALLOWED))
         assertThat(appWithChaos(Request(GET, "/")), hasStatus(NOT_FOUND))
         assertThat(appWithChaos(Request(POST, "/chaos/deactivate")), hasStatus(OK).and(hasBody(noChaos)))
         assertThat(appWithChaos(Request(GET, "/chaos/status")), hasBody(noChaos))
         assertThat(appWithChaos(Request(GET, "/")), hasStatus(OK))
-        assertThat(appWithChaos(Request(POST, "/chaos/activate/new").body("""
+        assertThat(
+            appWithChaos(
+                Request(POST, "/chaos/activate/new").body(
+                    """
                    [{
                        "type":"trigger",
                        "trigger": {
@@ -87,7 +89,10 @@ class ChaosEngineTest {
                            "type":"status",
                            "status":418
                        }
-                   }]""".trimIndent())), hasStatus(OK).and(hasBody(customChaos)))
+                   }]""".trimIndent()
+                )
+            ), hasStatus(OK).and(hasBody(customChaos))
+        )
         assertThat(appWithChaos(Request(GET, "/chaos/status")), hasBody(customChaos))
         assertThat(appWithChaos(Request(GET, "/")), hasStatus(I_M_A_TEAPOT))
         assertThat(appWithChaos(Request(POST, "/chaos/deactivate")), hasStatus(OK).and(hasBody(noChaos)))
@@ -110,7 +115,8 @@ class ChaosEngineTest {
 
     @Test
     fun `combines with other route blocks`() {
-        val app = routes("/{bib}/{bar}" bind GET to { Response(I_M_A_TEAPOT).body(it.path("bib")!! + it.path("bar")!!) })
+        val app =
+            routes("/{bib}/{bar}" bind GET to { Response(I_M_A_TEAPOT).body(it.path("bib")!! + it.path("bar")!!) })
 
         val appWithChaos = app.withChaosApi(controlsPath = "/context")